+++ R/cb_create_spss.r
@@ -19,61 +19,47 @@
 #'   or [tidyselect][dplyr_tidy_select] expressions), and missing values on the
 #'   right-hand side. If left-hand side is omitted, defaults to `tidyselect::everything()`.
 #'   See "Specifying user missing values" in [`cb_create()`] documentation for examples.
-#' @param .rmv_html Should HTML tags be removed from variable and value labels?
-#' @param .rmv_line_breaks Should line breaks be removed from variable and value 
-#' labels? If `TRUE`, line breaks will be replaced with `" / "`.
-#' @param .user_missing_conflict If labels passed to `.user_missing` conflicts with
-#'   a value label in `data`, which should be used?
 #' 
 #' @return
-#' An `"li_codebook"` object, consisting of (1) a tibble summarizing the passed
-#' dataset and (2) attributes containing the passed dataset (in several formats)
-#' and additional metadata. Specifically:
-#' - A tibble with columns:
-#'     - `name`: variable name
-#'     - `type`: optional column containing simplified variable type
-#'     - `class`: optional column containing class(es) of each variable
-#'     - `label_stem`: optional column containing variable label stems, if any variables 
-#'       are specified in `.split_var_labels`
-#'     - `label`: variable label
-#'     - `values`: values, with labels if applicable
-#'     - `user_missing`: optional column, depending on value of `.user_missing_col`,
-#'        showing user missing values, with labels if applicable
-#'     - `missing`: proportion missing
-#' - Attributes:
-#'     - Transformed versions of the passed dataset. See [`cb_get_data()`].
-#'     - Lookup tables and other metadata used internally.
+#' An `"li_codebook"` object, consisting of a tibble summarizing the passed
+#' dataset and attributes containing additional metadata. The tibble includes columns:
+#' - `name`: variable name
+#' - `type`: column containing simplified variable type
+#' - `class`: optional column containing class(es) of each variable
+#' - `label_stem`: optional column containing variable label stems, if any variables
+#'   are specified in `.split_var_labels`
+#' - `label`: variable label
+#' - `values`: values, with labels if applicable
+#' - `user_missing`: optional column showing user missing values, with labels 
+#'   if applicable. By default, this column is included only if user missings 
+#'   are specified for at least one variable. This behavior can be changed using 
+#'   the `user_missing_col` argument to `cb_create_options()`.
+#' - `missing`: proportion missing
 #'
 #' @export
 cb_create_spss <- function(data,
                            .user_missing = NULL,
                            .split_var_labels = NULL,
-                           .include_types = !.include_r_classes,
-                           .include_r_classes = FALSE,
-                           .rmv_html = TRUE,
-                           .rmv_line_breaks = TRUE,
-                           .user_missing_col = c("if_any", "yes", "no"),
-                           .user_missing_conflict = c("val_label", "missing_label"),
-                           .user_missing_incompatible = c("ignore", "warn", "error")
-                           ) {
+                           .options = cb_create_options()) {
+  check_options(.options)
   data |>
     cb_init() |>
     cb_clean_fields_spss(
-      rmv_html = .rmv_html, 
-      rmv_line_breaks = .rmv_line_breaks
+      rmv_html = .options$rmv_html, 
+      rmv_line_breaks = .options$rmv_line_breaks
     ) |> 
     cb_add_label_col_spss() |>
     cb_update_labels_spss(
       user_missing = .user_missing,
-      user_missing_conflict = .user_missing_conflict,
-      user_missing_incompatible = .user_missing_incompatible
+      user_missing_conflict = .options$user_missing_conflict,
+      user_missing_incompatible = .options$user_missing_incompatible
     ) |>
     cb_zap_data_spss() |>
     cb_add_dims() |>
-    cb_add_val_labels_col(user_missing_col = .user_missing_col) |>
+    cb_add_val_labels_col(user_missing_col = .options$user_missing_col) |>
     cb_add_type_col(
-      include_r_classes = .include_r_classes,
-      include_types = .include_types
+      include_r_classes = .options$include_r_classes,
+      include_types = .options$include_types
     ) |>
     cb_add_missing_col() |>
     cb_split_labels_col(split_var_labels = rlang::enexpr(.split_var_labels))